add optional callback method to notify evictions
Often it is beneficial to have notifications when an item is being evicted from the cache.
With this PR, the LRU cache can be created with an optional callback function which will
be called when an item is being evicted from the cache.

Example:

````
julia> using LRUCache

julia> CACHE = LRU{Int, Int}(3; callback=(x,y)->println("removed $x=>$y"))
       for i in 1:4
           CACHE[i] = i*10
       end
removed 1=>10
````